+++ src/Examples/ExampleTemplate.cs
@@ -1,14 +1,24 @@
 namespace Examples
 {
     using System;
-    using System.Collections.Generic;
     using System.Threading.Tasks;
-    using Stripe;
-    using Stripe.V2.Billing;
 
-    public class NewExample
+    /// <summary>
+    ///
+    /// This is a template for defining new examples.  It is not intended to be used directly.
+    ///
+    /// <describe what this example does>
+    ///
+    /// In this example, we:
+    ///   - <key step 1>
+    ///   - <key step 2
+    ///   - ...
+    ///
+    /// <describe assumptions about the user's stripe account, environment, or configuration;
+    /// or things to watch out for when running>
+    /// </summary>
+    public class ExampleTemplate
     {
-
         public static async Task Run()
         {
             var apiKey = "{{API_KEY}}";
